requirements.txt
fastapi
uvicorn
earthengine-api
rasterio
numpy
google-generativeai
python-dotenv
fpdf2
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
httpx
pystac-client
odc-stac
opencv-python-headless
shapely
pandas